Move WorkspaceRemovedEvent and WorkspaceCreatedEvent to che-core-api-workspace-shared; Move AppStatesPreferenceCleaner to che-core-api-user
